+++ src/components/Typography/__test__/Paragraph.test.tsx
@@ -1,10 +1,10 @@
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
-import { Typography } from '../Typography';
+import { Paragraph } from '../components';
 
 describe('Typography.Paragraph', () => {
   it('should render the paragraph text', () => {
-    const { container } = render(<Typography.Paragraph>This is a paragraph</Typography.Paragraph>);
+    const { container } = render(<Paragraph>This is a paragraph</Paragraph>);
     const paragraph = screen.getByText('This is a paragraph');
 
     expect(paragraph).toBeInTheDocument();